move unrelated tests out of test_pipeline.py
